+++ Matrix/Permutation.hh
@@ -33,22 +33,22 @@ using std::vector;
 class Permutation {
 public:
     /// Constructor
-    Permutation(unsigned int n);
+    Permutation(size_t n);
     /// Destructor
     ~Permutation() {}
     
     /// get size
-    unsigned int size() const { return data.size(); }
+    size_t size() const { return data.size(); }
     /// immutable element access
-    unsigned int operator[](unsigned int n) const { assert(n<size()); return data[n]; }
+    size_t operator[](size_t n) const { assert(n<size()); return data[n]; }
     /// mutable element access
-    unsigned int& operator[](unsigned int n) { assert(n<size()); return data[n]; }
+    size_t& operator[](size_t n) { assert(n<size()); return data[n]; }
     
     /// Set to permutation matrix needed by ShieldPhysics
     Permutation& nshuffle(int n);
     
     /// Permute two elements
-    void swap(unsigned int a, unsigned int b);
+    void swap(size_t a, size_t b);
     
     /// Return the inverse (=transpose) of the matrix
     Permutation inverse() const;
@@ -63,7 +63,7 @@ public:
     const Permutation operator*(const Permutation& p) const;
     
 private:
-    vector<unsigned int> data; ///< the permutation data
+    vector<size_t> data; ///< the permutation data
 };
 
 #endif
